// src/app/shell/interfaces/table-output.ts
export interface TableOutput {
  SeriesName: string;
  SeriesCode: string;
  SeriesVariableID: string;
  SeriesMap: string;
  SeriesTree: string;
  CategoryName: string;
  CategoryCode: string;
  CategoryVariableID: string;
  CategoryMap: string;
  CategoryTree: string;
  Score: number;
  Difference: number;
  NestedCategoryMap: string;
  SigKey: string;
  SignificanceSign: number;
}